+++ addons/compat_rnt/CfgAnimationSoundSources.hpp
