--- addons/compat_rnt/CfgMagazineGroups.hpp
@@ -1,6 +1,0 @@
-class ace_csw_groups {
-    CREATE_MAGAZINE_GROUP(Redd_Mg3_Belt_100);
-
-    CREATE_MAGAZINE_GROUP(150Rnd_762x51_Box);
-    CREATE_MAGAZINE_GROUP(150Rnd_762x51_Box_Tracer);
-};
