--- addons/compat_rnt/CfgMagazineWells.hpp
@@ -1,8 +1,0 @@
-class CfgMagazineWells {
-    class GVAR(mg3_762x51) {
-        MG_Magazines[] = {
-            "150Rnd_762x51_Box",
-            "150Rnd_762x51_Box_Tracer"
-        };
-    };
-};
