--- addons/compat_rnt/XEH_PREP.hpp
@@ -1,5 +1,0 @@
-PREP(canAdjustHeight);
-PREP(canMountMG);
-PREP(disassemble);
-PREP(mountMG);
-PREP(pickupTripod);
