--- addons/compat_rnt/XEH_preInit.sqf
@@ -1,12 +1,0 @@
-#include "script_component.hpp"
-
-ADDON = false;
-
-PREP_RECOMPILE_START;
-#include "XEH_PREP.hpp"
-PREP_RECOMPILE_END;
-
-ADDON = true;
-
-
-#include "initSettings.inc.sqf"
