+++ addons/compat_rnt/compat_rnt_luchs/config.cpp
@@ -1,0 +1,20 @@
+#include "script_component.hpp"
+
+class CfgPatches {
+    class SUBADDON {
+        authors[] = {"Mishkar", "Andx"};
+        name = COMPONENT_NAME;
+        addonRootClass = QUOTE(ADDON);
+        requiredVersion = REQUIRED_VERSION;
+        requiredAddons[] = {
+            "redd_tank_fuchs_1a4"
+        };
+        units[] = {};
+        weapons[] = {};
+        VERSION_CONFIG;
+
+        skipWhenMissingDependencies = 1;
+    };
+};
+
+#include "CfgVehicles.hpp"
