+++ addons/compat_rnt/compat_rnt_luchs/script_component.hpp
@@ -1,0 +1,3 @@
+#define COMPONENT fuchs
+#define COMPONENT_BEAUTIFIED Fuchs
+#include "..\script_component.hpp"
