--- addons/compat_rnt/readme.md
@@ -1,17 +1,0 @@
-# Compat RnT
-
-## Fixes
-
-- repariert ACE-Aktionen an MG3 und GMW
-
-## Geändert
-
-- ACRE Racks, BV und Infanterietelephon angepasst
-
-## Abhängigkeiten
-
-- Redd'n'Tank <https://steamcommunity.com/sharedfiles/filedetails/?id=1200127537>
-
-## Maintainers
-
-- Andx
